requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
openai>=1.12.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.0.0
pytesseract>=0.3.10

# Vector storage
chromadb>=0.4.22

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.5.2

# Hebrew text support
python-bidi>=0.4.2
arabic-reshaper>=3.0.0
